+++ app/models/planning_element_type_color.rb
@@ -17,7 +17,7 @@ class PlanningElementTypeColor < ActiveRecord::Base
   acts_as_list
   default_scope :order => 'position ASC'
 
-  has_many :planning_element_types, :class_name  => 'PlanningElementType',
+  has_many :planning_element_types, :class_name  => 'Type',
                                     :foreign_key => 'color_id',
                                     :dependent   => :nullify