+++ app/models/work_package.rb
@@ -32,14 +32,15 @@ class WorkPackage < ActiveRecord::Base
   belongs_to :priority, :class_name => 'IssuePriority', :foreign_key => 'priority_id'
   belongs_to :category, :class_name => 'IssueCategory', :foreign_key => 'category_id'
 
-  belongs_to :planning_element_type,   :class_name  => "PlanningElementType",
-                                       :foreign_key => 'planning_element_type_id'
   belongs_to :planning_element_status, :class_name  => "PlanningElementStatus",
                                        :foreign_key => 'planning_element_status_id'
 
   has_many :time_entries, :dependent => :delete_all
   has_many :relations_from, :class_name => 'IssueRelation', :foreign_key => 'issue_from_id', :dependent => :delete_all
   has_many :relations_to, :class_name => 'IssueRelation', :foreign_key => 'issue_to_id', :dependent => :delete_all
+  has_and_belongs_to_many :changesets,
+                          :order => "#{Changeset.table_name}.committed_on ASC, #{Changeset.table_name}.id ASC"
+
 
   scope :recently_updated, :order => "#{WorkPackage.table_name}.updated_at DESC"
   scope :visible, lambda {|*args| { :include => :project,
@@ -107,7 +108,6 @@ class WorkPackage < ActiveRecord::Base
                                                     :status_id, :type_id,
                                                     :assigned_to_id, :priority_id,
                                                     :category_id, :fixed_version_id,
-                                                    :planning_element_type_id,
                                                     :planning_element_status_id,
                                                     :author_id, :responsible_id
   register_on_journal_formatter :datetime,          :start_date, :due_date, :deleted_at
@@ -294,11 +294,7 @@ class WorkPackage < ActiveRecord::Base
   end
 
   def kind
-    if self.is_a? Issue
-      return type
-    elsif self.is_a? PlanningElement
-      return planning_element_type
-    end
+    return type
   end
 
   def to_s
@@ -308,6 +304,11 @@ class WorkPackage < ActiveRecord::Base
   # Return true if the work_package is closed, otherwise false
   def closed?
     self.status.nil? || self.status.is_closed?
+  end
+
+  # Returns true if the work_package is overdue
+  def overdue?
+    !due_date.nil? && (due_date < Date.today) && !closed?
   end
 
   # TODO: move into Business Object and rename to update
@@ -328,14 +329,23 @@ class WorkPackage < ActiveRecord::Base
   end
 
   def is_milestone?
-    planning_element_type && planning_element_type.is_milestone?
+    type && type.is_milestone?
   end
 
-  # This is a dummy implementation that is currently overwritten
-  # by issue
-  # Adapt once tracker/type is migrated
-  def new_statuses_allowed_to(user, include_default = false)
-    IssueStatus.all
+  # Returns an array of status that user is able to apply
+  def new_statuses_allowed_to(user, include_default=false)
+    return [] if status.nil?
+
+    statuses = status.find_new_statuses_allowed_to(
+      user.roles_for_project(project),
+      type,
+      author == user,
+      assigned_to_id_changed? ? assigned_to_id_was == user.id : assigned_to_id == user.id
+      )
+    statuses << status unless statuses.empty?
+    statuses << IssueStatus.default if include_default
+    statuses = statuses.uniq.sort
+    blocked? ? statuses.reject {|s| s.is_closed?} : statuses
   end
 
   def self.use_status_for_done_ratio?