+++ features/step_definitions/repository_steps.rb
@@ -9,9 +9,16 @@
 # See doc/COPYRIGHT.rdoc for more details.
 #++
 
-FactoryGirl.define do
-  factory(:default_planning_element_type, :class => DefaultPlanningElementType) do
-    project_type          { |e| e.association(:project_type) }
-    planning_element_type { |e| e.association(:planning_element_type) }
-  end
+Given(/^the project "(.*?)" has a repository$/) do |project_name|
+
+  project = Project.find(project_name)
+
+  repo = FactoryGirl.build(:repository,
+                           :project => project)
+
+  Setting.enabled_scm = Setting.enabled_scm << repo.scm_name
+
+  repo.save!
 end
+
+