+++ features/step_definitions/timelines_given_steps.rb
@@ -30,18 +30,6 @@ Given /^I am working in the [tT]imeline "([^"]*)" of the project called "([^"]*)
   @timeline_name = timeline_name
 end
 
-Given /^there are the following planning element types:$/ do |table|
-  # Color is not handled in a sensible way. We need some extra logic to match
-  # a color name to an id, so that it is possible to assign a certain color to
-  # planning element types. This should be added once it is needed.
-  #
-  table.map_headers! { |header| header.underscore.gsub(' ', '_') }
-
-  table.hashes.each do |type_attributes|
-    FactoryGirl.create(:planning_element_type, type_attributes)
-  end
-end
-
 Given /^there are the following planning element statuses:$/ do |table|
   table.map_headers! { |header| header.underscore.gsub(' ', '_') }
 
@@ -66,17 +54,6 @@ Given /^there are the following project types:$/ do |table|
   end
 end
 
-Given /^the following planning element types are default for projects of type "([^"]*)"$/ do |project_type_name, pe_type_names|
-  project_type = ProjectType.find_by_name!(project_type_name)
-
-  pe_type_names = pe_type_names.raw.flatten
-  pe_type_names.each do |pe_type_name|
-    FactoryGirl.create(:default_planning_element_type,
-                   :project_type_id          => project_type.id,
-                   :planning_element_type_id => PlanningElementType.find_by_name!(pe_type_name).id)
-  end
-end
-
 Given /^there is a scenario "([^"]*)" in project "([^"]*)"$/ do |scenario_name, project_name|
   FactoryGirl.create(:scenario, :name => scenario_name, :project_id => Project.find_by_name!(project_name).id)
 end
@@ -96,15 +73,6 @@ end
 Given /^I delete the scenario "([^"]*)"$/ do |scenario_name|
   scenario = Scenario.find_by_name!(scenario_name)
   scenario.destroy
-end
-
-# Using our own project creation step to make sure, that we may initially assign
-# a project type.
-#
-Given /^there is a project named "([^"]*)" of type "([^"]*)"$/ do |name, project_type_name|
-  FactoryGirl.create(:project,
-                 :name                      => name,
-                 :project_type_id => ProjectType.find_by_name!(project_type_name).id)
 end
 
 Given /^there are the following projects of type "([^"]*)":$/ do |project_type_name, table|
@@ -142,3 +110,15 @@ Given /^there is a timeline "([^"]*)" for project "([^"]*)"$/ do |timeline_name,
   timeline.save!
 end
 
+Given /^the following types are enabled for projects of type "(.*?)"$/ do |project_type_name, type_name_table|
+  project_type = ProjectType.find_by_name(project_type_name)
+  projects = Project.where(:project_type_id => project_type.id)
+  types = type_name_table.raw.flatten.map do |type_name|
+    Type.find_by_name(type_name) || Factory.create(:type, :name => type_name)
+  end
+
+  projects.each do |project|
+    project.types = types
+    project.save
+  end
+end