+++ lib/redmine.rb
@@ -53,7 +53,7 @@ end
 Redmine::AccessControl.map do |map|
   map.permission :view_project,
                  {
-                   :planning_element_types => [:index, :show],
+                   :types => [:index, :show],
                    :projects => [:show],
                    :scenarios => [:index, :show],
                    :projects => [:show],
@@ -271,9 +271,6 @@ Redmine::MenuManager.map :admin_menu do |menu|
   menu.push :colors,
             {:controller => '/planning_element_type_colors', :action => 'index'},
             {:caption    => :'timelines.admin_menu.colors' }
-  menu.push :planning_element_types,
-            {:controller => '/planning_element_types', :action => 'index'},
-            {:caption    => :'timelines.admin_menu.planning_element_types' }
   menu.push :project_types,
             {:controller => '/project_types', :action => 'index'},
             {:caption    => :'timelines.admin_menu.project_types' }